Move functions into utils script
