+++ tests/test_file_data.py
@@ -6092,6 +6092,8 @@ class DiagnosticsOnKnownFilledPostsData:
         ),
     ]
 
+@dataclass
+class DiagnosticsUtilsData:
     filter_to_known_values_rows = [
         (
             "loc 1",
